+++ SistemaVentas/src/productos/producto.controller.js
@@ -1,8 +1,14 @@
 'use strict'
 import Producto from './producto.modelo.js'
+import Categoria from '../categorias/categorias.model.js'
+import Usuario from '../usuario/usuario.modelo.js'
+import Carrito from '../Carrito/carrito.model.js'
 
 export const nuevo = async (req, res) => {
     try {
+        let {uid}= req.user
+        let usuarioADMIN = await Usuario.findOne({_id: uid})
+        if(usuarioADMIN.role != 'ADMIN') return res.status(401).send({message: 'Debes ser admin para realizar eeste cambio'}) 
         let data = req.body
         let product = new Producto(data)
         await product.save()
@@ -16,12 +22,27 @@ export const nuevo = async (req, res) => {
 export const buscar = async (req, res) => {
     try {
         let { id } = req.params
-        let product = await Producto.findOne({ _id: id })
-        return res.status(200).send({ product })
+        let producto = await Producto.findOne({ _id: id })
+        return res.status(200).send({ producto })
 
     } catch (err) {
         console.error(err);
         return res.status(404).send({ message: 'El producto no se encontrado' })
+    }
+}
+
+export const BuscarxCategoria = async (req, res) => {
+    try {
+        let data = req.body
+        let { id } = req.params
+        data.categoria = id
+        let categoria = await Categoria.findOne({ _id: id })
+        if (!categoria) return res.status(404).send({ message: 'La categoria que buscas no esta disponible' })
+        let producto = await Producto.find({categoria})
+        if (!producto) return res.status(404).send({ message: 'El producto que buscas no existe' })
+        return res.send({ message:`Productos de la categoria ${data.categoria}:`, producto })
+    } catch (err) {
+        console.error(err);
     }
 }
 
@@ -36,6 +57,9 @@ export const listar = async (req, res) => {
 
 export const actualizar = async (req, res) => {
     try {
+        let {uid}= req.user
+        let usuarioADMIN = await Usuario.findOne({_id: uid})
+        if(usuarioADMIN.role != 'ADMIN') return res.status(401).send({message: 'Debes ser admin para realizar eeste cambio'}) 
         let { id } = req.params
         let data = req.body
         let actualizar = await Producto.findOneAndUpdate({ _id: id }, data, { new: true })
@@ -49,6 +73,10 @@ export const actualizar = async (req, res) => {
 
 export const eliminar = async (req, res) => {
     try {
+        let {uid}= req.user
+        let usuarioADMIN = await Usuario.findOne({_id: uid})
+        if(usuarioADMIN.role != 'ADMIN') return res.status(401).send({message: 'Debes ser admin para realizar eeste cambio'}) 
+       
         let { id } = req.params
         let eliminar = await Producto.findOneAndDelete({ _id: id })
         if (!eliminar) return res.status(404).send({ message: 'Producto no encontrado para  poder eliminarlo' })
@@ -60,22 +88,27 @@ export const eliminar = async (req, res) => {
 
 export const agotados = async (req, res) => {
     try {
-
+        let {uid}= req.user
+        let usuarioADMIN = await Usuario.findOne({_id: uid})
+        if(usuarioADMIN.role != 'ADMIN') return res.status(401).send({message: 'Debes ser admin para realizar eeste cambio'}) 
         let agotados = await Producto.find({ stock: 0 })
-        if(agotados.length == 0) return res.status(404).send({message:'No hay productos agotados'})
-        return res.send({message:'Estos son los productos que estn agotados', agotados })
-       
+        if (!agotados.length === 0) return res.status(404).send({ message: 'No hay productos agotados' })
+        return res.send({ message: 'Estos son los productos que estn agotados', agotados })
     } catch (err) {
         console.error(err);
         return res.status(500).send({ message: 'Error en la base de datos' })
     }
-
 }
 
-export const masVendidos =async(req, res)=>{
-    try{
 
-    }catch(err){
+
+
+export const masVendidos = async (req, res) => {
+    try {
+
+    } catch (err) {
         console.error(err);
     }
 }
+
+
